OAuth, Basic business permissions, Password bcrypted

+++ project/app/routers/business_router.py
@@ -1,12 +1,12 @@
-from fastapi import Depends, APIRouter, HTTPException, Path, Query
+from fastapi import Depends, APIRouter, HTTPException, Path, Query, status
 from typing import Annotated
 from sqlmodel import Session
 from ..database import get_db
 from ..models import Business
 from ..services.business_service import BusinessService
 from ..schemas.business_schema import BusinessUpdate
-from ..schemas.login_schema import Login
-
+from ..helpers.security.security_helper import UserSecurityHelper
+from ..models import User
 
 
 router = APIRouter(prefix = "/business", tags = ['business'])
@@ -28,6 +28,28 @@ async def get_all(
         dict: Respuesta del servicio.
     """
     return await BusinessService.get_all(db, offset, limit)
+
+
+@router.get("/me")
+async def get_me(
+    current_user: Annotated[User, Depends(UserSecurityHelper.get_current)],
+    offset: Annotated[int, Query(title = "The page of business we want to get")] = 0,
+    limit:  Annotated[int, Query(title = "The number of businesses we want to get per page")] = 10,
+    db: Session = Depends(get_db),
+    ):
+    """Get by id: Dado un id, regresa su negicio correspondiente 
+
+    Args:
+        id (Annotated[int, Path, optional): Id de entidad. Defaults to "ID of the business we want to find")].
+        db (Session, optional): Motor de base de datos. Defaults to Depends(get_db).
+
+    Raises:
+        HTTPException: Error de HTTP, da una respuesta con código.
+
+    Returns:
+        dict: Respuesta del servicio.
+    """
+    return await BusinessService.get_all(db, offset, limit, { 'holder_id': current_user.id })
 
 @router.get("/{id}")
 async def get_by_id(
@@ -51,8 +73,10 @@ async def get_by_id(
         raise HTTPException(status_code = 404, detail = "business not found")
     return user
 
+
 @router.post("/", status_code = 201)
 async def create(
+    current_user: Annotated[User, Depends(UserSecurityHelper.get_current)],
     business:Business,
     db: Session = Depends(get_db),
     ):
@@ -68,37 +92,18 @@ async def create(
     Returns:
         dict: Respuesta del servicio.
     """
+    business.holder_id = current_user.id
     new_business = await BusinessService.create(db, business)
     if not new_business:
         raise HTTPException(status_code = 500, detail = "couldn't create Business")
     return new_business
 
-@router.post('/auth')
-async def login(
-    login: Login,
-    db: Session = Depends(get_db),
-    ):
-    """Login: dado un email y su contraseña, autentifica y regresa una empresa.
-
-    Args:
-        login (Login): Objeto con email y contraseña.
-        db (Session, optional): Motor de base de datos. Defaults to Depends(get_db).
-
-    Raises:
-        HTTPException: Error de HTTP, da una respuesta con código.
-
-    Returns:
-        dict: Respuesta del servicio.
-    """
-    user = await BusinessService.login(db, login)
-    if not user:
-        raise HTTPException(status_code = 403, detail = "invalid credentials")
-    return user
 
 @router.put("/{id}")
 async def update(
+    current_user: Annotated[User, Depends(UserSecurityHelper.get_current)],
     id: Annotated[int, Path(title="ID of the business we want to update")],
-    business:BusinessUpdate,
+    businessUp:BusinessUpdate,
     db: Session = Depends(get_db),
     ):
     """Update: Actualiza un registro según datos mutables y su id.
@@ -114,13 +119,18 @@ async def update(
     Returns:
         dict: Respuesta del servicio.
     """
-    new_business = await BusinessService.update(db, business, id)
+    business = await BusinessService.get_by_id(db, id)
+    if business.holder_id != current_user.id:
+        raise HTTPException(status_code = status.HTTP_403_FORBIDDEN , detail = "you don't have permission for this action")
+    new_business = await BusinessService.update(db, businessUp, id)
     if not new_business:
         raise HTTPException(status_code = 500, detail = "couldn't update Business")
     return new_business
 
+
 @router.delete("/{id}")
 async def delete(
+    current_user: Annotated[User, Depends(UserSecurityHelper.get_current)],
     id: Annotated[int, Path(title="ID of the business we want to deactivate")],
     db: Session = Depends(get_db),
     ):
@@ -136,6 +146,9 @@ async def delete(
     Returns:
         dict: Respuesta del servicio.
     """
+    business = await BusinessService.get_by_id(db, id)
+    if business.holder_id != current_user.id:
+        raise HTTPException(status_code = status.HTTP_403_FORBIDDEN , detail = "you don't have permission for this action")
     new_business = await BusinessService.delete(db, id)
     if not new_business:
         raise HTTPException(status_code = 500, detail = "couldn't delete Business")
